requirements.txt
# ── Core Pipeline ──────────────────────────────────────────────
langgraph>=0.2.0
langchain-google-genai>=2.0.0
google-genai>=1.0.0
sentence-transformers>=2.7.0

# ── PDF Parsing ───────────────────────────────────────────────
docling>=2.0.0

# ── Chunking ──────────────────────────────────────────────────
llama-index-core>=0.11.0
llama-index-embeddings-google>=0.3.0

# ── Vector DB ─────────────────────────────────────────────────
pinecone>=5.0.0

# ── Observability ─────────────────────────────────────────────
mlflow>=2.15.0
langsmith>=0.1.0

# ── Data & Utilities ─────────────────────────────────────────
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0

# ── Report Generation ────────────────────────────────────────
reportlab>=4.0.0

# ── Ground Truth / Z-Score ────────────────────────────────────
yfinance>=0.2.30

# ── Dev ───────────────────────────────────────────────────────
pytest>=8.0.0
ruff>=0.4.0
